chore: moved socket to core as it is not really specific to http
